infrastructure/requirements.txt
# CDK Requirements for AgentCore Weather Tools Infrastructure
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# Additional dependencies for deployment automation
boto3>=1.26.0
botocore>=1.29.0